motion_3.c: Blinkers, motion_4.c: Sound

+++ Lab1/math_functions.h
@@ -1,12 +1,12 @@
 // math.h
 
-int min(int value1, int value2)
+float min(float value1, float value2)
 {
    return (value2 > value1 ? value1 : value2);
 }
 
 
-int max(int value1, int value2)
+float max(float value1, float value2)
 {
    return (value2 < value1 ? value1 : value2);
 }
